Fix Issue 14
Issue: https://github.com/srct/schedules/issues/14

Fixed issue by doing the following:
When processing a CRN, add a hash to the url so that when the course page is loaded, the section with the CRN will be jumped to. (app/controllers/search_controller.rb)

Additionally, with CSS, highlight the target course with a green color (#afa) to signify that it is the desired course. (app/assets/stylesheets/application.scss)

Turbolink messes with the behaviour of anchors, so when the page is loaded, a script makes sure that the hash is acted upon. (app/views/courses/show.html)