Added instructor rating in search and removed the individual instructors' course lists
The rating is rendered in mostly the same way it is on the instructor page. It doesn't display the number of responses and it formats the number to always have two decimal places.

The drop-down each instructor had is still there, and it has had some upgrades since the last commit, but I deemed that the information in the drop-down was redundant since it could be accessed by just clicking on the instructor's name, as that information can be found on their individual pages.

However, I made it so the search for courses taught by the instructor is never made, so those elements are never actually added to the page. If this feature is something anyone wants to experiment with, I've left more info in the app/views/search/index.html.erb file.